Relax LNURLP status checks (#6)
* Relax LNURLP status checks

Some LNURLP implementations (such as the LNBits 0.12.x on demo.lnpiggy.com)
don't send a "status" field in their initial response, nor in their invoice reponse.

This was causing getInvoiceFromLNAddr() to fail,
even though it works fine with other tested clients.

Instead of failing if the status is not present or if it's not OK,
this commit only fails if the status is not OK,
but continues if there is no explicit status field set.

* Use NostrString_equals() on NostrString