Nip04::decrypt - Allocate memory on the heap instead of the stack (#9)
* Allocate memory on the heap instead of the stack

When receiving big responses, such as for list_transactions, the memory
constraints of the ESP32 really become apparent, especially when using the
pre-built esp-idf from the Arduino IDE's board manager, with its fixed configuration.

The stack memory size is particularly constaining. It seems to have around ~4KB free when
the code ends up in the loop() function, so there's really no room to allocate big arrays there.

A typical transaction from a list_transactions reply is around 4KB on its own because NWC is
a pretty verbose protocol (JSON encapsulation, encryption, base64 encoding, hex encoding)
and it was causing a crash due to all the stack memory being used up.

This modification allocates the memory on the heap using malloc(), which fixes the issue.

The heap memory itself is also pretty limited, and there are probably other places that would need
to be rewritten to improve this further, or to be able to handle arbitrarily large transactions.

* Use NostrString for portability