Add new Arduino esp32 example and prepare for next release (#16)
* Add new extensive Arduino ESP32 NWC example

This example is picked up by the Arduino IDE because it has the .ino
extension, so it will be listed under "Examples" - "Examples from Custom Libraries"
- "Nostrduino" - ArduinoESP32TestNWC.

It has the necessary imports to make it compile (tested with) Arduino IDE 1.8.3

It's dependent on the next release of Nostrduino (1.3.6) because that one has
subscribeNotifications() to subscribe to payment notifications.

* Prepare for 1.3.6 release

A new release is needed to add support for subscribeNotifications
that is used in the new Arduino example.